smartQuizzer/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-login==0.6.3
werkzeug==2.3.7
google-generativeai==0.3.0
python-dotenv==1.0.0
requests==2.31.0